refactor: created attribute router trait + split AbstractRouter
