+++ test/Asset/AttributeRouteRouter.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * BitFrame Framework (https://www.bitframephp.com)
+ *
+ * @author    Daniyal Hamid
+ * @copyright Copyright (c) 2017-2022 Daniyal Hamid (https://designcise.com)
+ * @license   https://bitframephp.com/about/license MIT License
+ */
+
+declare(strict_types=1);
+
+namespace BitFrame\Test\Asset;
+
+use BitFrame\Router\AttributeRouteTrait;
+
+class AttributeRouteRouter extends MultiRouteRouter
+{
+    use AttributeRouteTrait;
+}
